Refactor widget notification handling
